// ecommerce/src/store/cart/types.ts
import { IProduct } from "../../types";

export enum actionTypes {
  CART_ADD_ITEM = 'CART_ADD_ITEM',
  CART_REMOVE_ITEM = 'CART_REMOVE_ITEM',
}

interface ICartAddItemPayLoad {
  product: IProduct
  count: number 
}
interface ICartAddItem {
  type: actionTypes.CART_ADD_ITEM
  payload: ICartAddItemPayLoad
}

interface ICartRemoveItem {
  type: actionTypes.CART_REMOVE_ITEM
  payload: string
}

export type TypeActionCart = ICartAddItem | ICartRemoveItem